Update to current `cryptobox`.
The internal use of `Arc`s in `CBox` requires us to either assert that
any `CBox` or `CBoxSession` is unwind-safe or to no longer catch stack
unwinding and instead choose another panic strategy such as aborting.
This merge requests implements the first approach. Client code which
encounters a `CBOX_PANIC` result must no longer use the `CBox` or any
of it's `CBoxSession`s.